move key and value transformation functions from metadata_file to utils
